+++ ios/Classes/Md5FileChecksumPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface Md5FileChecksumPlugin : NSObject<FlutterPlugin>
-@end