+++ scripts/generate-pigeon-files.bash
@@ -6,8 +6,7 @@
 dart run pigeon \
   --input pigeon/interface.dart \
   --dart_out lib/src/interface.dart \
-  --objc_header_out ios/Classes/pigeon.h \
-  --objc_source_out ios/Classes/pigeon.m \
+  --swift_out ios/md5_file_checksum/Sources/md5_file_checksum/Pigeon.swift \
   --java_out ./android/src/main/java/dev/flutter/pigeon/Pigeon.java \
   --java_package "dev.flutter.pigeon"
 dart format lib/